fix(EPIC-001): resolve all 19 PR review comments — eliminate u128 truncation
Category A — Sled storage boundary:
- Widen backfill_token_balances_from_contract to accept u128 entries
- Widen get_token_supply/put_token_supply to u128 in storage trait + sled
- Backward-compat sled reads: accept both 8-byte (legacy u64) and 16-byte u128
- Remove all as-u64 casts in seed_map and genesis balance persistence

Category B — API response types:
- WalletInfo.available_balance/total_balance → u128
- TokenInfoResponse/TokenListItem supply fields → u128
- ExplorerStatsResponse.total_supply → u128

Category C — Business logic:
- Voting power divisor: 1e8 → TOKEN_SCALE_18 (1e18) — LOGIC BUG fix
- get_dao_treasury_balance returns u128 (was u64)
- get_circulating_sov_supply returns u128 (was u64)
- can_mint uses checked_add to prevent u128 overflow
- Disbursement.token_burned → u128
- backup_recovery: movable_balance_total → u128, safe try_from at mint boundary
- srv_ops: safe u64::try_from for SRV circulating_supply

Refs: #2093, #2094

+++ lib-blockchain/src/execution/tx_apply.rs
@@ -173,12 +173,12 @@ impl<'a> StateMutator<'a> {
         let _burn_amount = if let Ok(Some(contract)) = self.store.get_token_contract(token) {
             if contract.is_deflationary && contract.burn_rate > 0 {
                 // burn_rate is in basis points (1/10000)
-                let burn = (amount * contract.burn_rate as u128 / 10_000) as u64;
+                let burn = amount * contract.burn_rate as u128 / 10_000;
                 if burn > 0 {
                     // Debit sender the full amount
                     self.debit_token(token, from, amount)?;
                     // Credit receiver the amount minus burn
-                    self.credit_token(token, to, amount.saturating_sub(burn as u128))?;
+                    self.credit_token(token, to, amount.saturating_sub(burn))?;
                     // Reduce total supply
                     if let Ok(Some(supply)) = self.store.get_token_supply(token) {
                         self.store
@@ -572,14 +572,14 @@ impl<'a> StateMutator<'a> {
     }
 
     /// Read token supply without mutating state.
-    pub fn get_token_supply(&self, token: &TokenId) -> TxApplyResult<Option<u64>> {
+    pub fn get_token_supply(&self, token: &TokenId) -> TxApplyResult<Option<u128>> {
         self.store
             .get_token_supply(token)
             .map_err(|e| TxApplyError::Storage(e.to_string()))
     }
 
     /// Persist token supply in canonical state storage.
-    pub fn put_token_supply(&self, token: &TokenId, supply: u64) -> TxApplyResult<()> {
+    pub fn put_token_supply(&self, token: &TokenId, supply: u128) -> TxApplyResult<()> {
         self.store
             .put_token_supply(token, supply)
             .map_err(|e| TxApplyError::Storage(e.to_string()))
